Add check for missing rule result operands and default for env
This adds a check that every policy rule explicitly says how to update any
metadata sets that the opgroup allows.  There is a special case for "env"
which may be left off if it is meant to stay the same.  I also added new
syntax for rule results - you can write just "allow" for the result if env
is the only result operand, or "allow with foo = bar, baz = ..." if you
want updates.  The old syntax is still supported.